Создание презентера и перенос метода convert
