+++ MovieQuiz/Presentation/AlertPresenter.swift
@@ -26,7 +26,7 @@ final class AlertPresenter {
             model.completion?()
         }
         alert.addAction(action)
-        alert.view.accessibilityIdentifier = "endGameAlert"
+        alert.view.accessibilityIdentifier = "AlertPresenter"
         viewController?.present(alert, animated: true, completion: nil)
         
     }
